in progress: improve code improve  Viewmodel RegisterAttendanceViewModel.kt colors function parse Qr valid error custom messages
