+++ app/src/main/AndroidManifest.xml
@@ -10,14 +10,14 @@
     <application
         android:allowBackup="true"
         android:usesCleartextTraffic="true"
-        android:name=".ui.BaseApplication"
+        android:name=".presentation.BaseApplication"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name_build"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.QRunlockclientapp">
         <activity
-            android:name=".ui.scanner.QrScanActivity"
+            android:name=".presentation.scanner.QrScanActivity"
             android:exported="false"
 
             android:theme="@style/FullScreen" />
@@ -25,7 +25,7 @@
             android:name=".TicketResultActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
