+++ app/src/main/java/com/rogergcc/qrunlockclientapp/data/network/QrEventClient.kt
@@ -2,24 +2,28 @@ package com.rogergcc.qrunlockclientapp.data.network
 
 import android.content.res.Resources
 import com.google.gson.Gson
-import com.rogergcc.qrunlockclientapp.data.model.AttedancesRows
-import com.rogergcc.qrunlockclientapp.data.response.RegisterAttendanceResponse
-import com.rogergcc.qrunlockclientapp.ui.helper.TimberAppLogger
+import com.rogergcc.qrunlockclientapp.data.network.request.UserAttendanceRequest
+import com.rogergcc.qrunlockclientapp.data.network.response.AttedancesRows
+import com.rogergcc.qrunlockclientapp.data.network.response.RegisterAttendanceResponse
+import com.rogergcc.qrunlockclientapp.presentation.extensions.fromJson
+import com.rogergcc.qrunlockclientapp.presentation.helper.TimberAppLogger
+import com.rogergcc.qrunlockclientapp.presentation.scanner.QrScanImage
 import retrofit2.Response
 import javax.inject.Inject
 
 class QrEventClient @Inject constructor(
-    private val apiService: QrEventsApi) {
+    private val apiService: QrEventsApi,
+) {
 
     suspend fun verifyRegisterUserAttendance(): List<AttedancesRows> {
-           return try{
-                val response = apiService.getVerifyUserAttendance()
-                response.body()?.body?.attedancesRows ?: emptyList()
-            }catch (e: QrEventInterceptor.NoInternetException){
+        return try {
+            val response = apiService.getVerifyUserAttendance()
+            response.body()?.body?.attedancesRows ?: emptyList()
+        } catch (e: QrEventInterceptor.NoInternetException) {
 //               Log.d(QrEventClient::class.java.simpleName, e.message, e)
-               TimberAppLogger.e("[QrEventClient] (verifyRegisterUserAttendace) ${e.message}")
-               emptyList()
-            }
+            TimberAppLogger.e("[QrEventClient] (verifyRegisterUserAttendace) ${e.message}")
+            emptyList()
+        }
     }
 
     suspend fun registerAttendanceRecord(userCode: String): RegisterAttendanceResponse {
@@ -27,34 +31,30 @@ class QrEventClient @Inject constructor(
             TimberAppLogger.d("registerAttendanceRecord: userCode $userCode")
 
             val response = apiService.registerAttendanceRecord(UserAttendanceRequest(userCode, 1))
-
+            TimberAppLogger.d("Raw response: $response")
             response.errorBody()?.use {
                 val errorBodyString = it.string()
-                val errorResponse = Gson().fromJson(errorBodyString, RegisterAttendanceResponse::class.java)
+
+//                val errorResponse = Gson().fromJson(errorBodyString, RegisterAttendanceResponse::class.java)
+                val errorResponse : RegisterAttendanceResponse = Gson().fromJson(errorBodyString)
                 TimberAppLogger.e("[QrEventClient] (registerAttendanceRecord) Error response: $errorResponse")
-                throw Exception(errorResponse.message)
+//                throw Exception(errorResponse.message)
+                throw handleHttpException(response)
+            }
+            if (!response.isSuccessful) {
+                TimberAppLogger.e("Unsuccessful response: ${response.code()} - ${response.message()}")
+                throw handleHttpException(response)
             }
 
-            TimberAppLogger.d("Raw response: $response")
-
-            // Check if the request was successful (status code 2xx)
-            if (response.isSuccessful) {
-                val responseBody = response.body() as RegisterAttendanceResponse
-                TimberAppLogger.i("Response body: $responseBody")
-                return responseBody
-            }
-
-            // Si no es exitoso, se llegará a este punto y se lanzará la excepción
-            TimberAppLogger.e("Unsuccessful response: ${response.code()} - ${response.message()}")
-            throw handleHttpException(response)
+            val responseBody = response.body() as RegisterAttendanceResponse
+            TimberAppLogger.i("Response body: $responseBody")
+            return responseBody
         } catch (e: Exception) {
-            // Handle other non-specific exceptions
-            TimberAppLogger.e("[QrEventClient] (registerAttendanceRecord) ex $e")
             TimberAppLogger.e("[QrEventClient] (registerAttendanceRecord) ex.message ${e.message}")
             throw Exception(e.message)
         }
-
     }
+
     private fun handleHttpException(response: Response<*>): Exception {
         return when (response.code()) {
             404 -> {
