+++ app/src/main/java/com/rogergcc/qrunlockclientapp/data/network/QrEventsApi.kt
@@ -1,7 +1,8 @@
 package com.rogergcc.qrunlockclientapp.data.network
 
-import com.rogergcc.qrunlockclientapp.data.model.AttendanceResponse
-import com.rogergcc.qrunlockclientapp.data.response.RegisterAttendanceResponse
+import com.rogergcc.qrunlockclientapp.data.network.response.AttendanceResponse
+import com.rogergcc.qrunlockclientapp.data.network.request.UserAttendanceRequest
+import com.rogergcc.qrunlockclientapp.data.network.response.RegisterAttendanceResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
