--- app/src/main/java/com/rogergcc/qrunlockclientapp/data/network/Resource.kt
@@ -1,21 +1,0 @@
-package com.rogergcc.qrunlockclientapp.data.network
-
-import okhttp3.ResponseBody
-
-
-/**
- * Created on marzo.
- * year 2024 .
- */
-sealed class Resource<out T> {
-
- data class Success<out T>(val value: T) : Resource<T>()
-
- data class Failure(
-  val isNetworkError: Boolean,
-  val errorCode: Int?,
-  val errorBody: ResponseBody?
- ) : Resource<Nothing>()
-
- object Loading : Resource<Nothing>()
-}
