+++ app/src/main/java/com/rogergcc/qrunlockclientapp/data/network/SafeApiCall.kt
@@ -1,5 +1,6 @@
 package com.rogergcc.qrunlockclientapp.data.network
 
+import com.rogergcc.qrunlockclientapp.common.Resource
 import retrofit2.HttpException
