+++ app/src/main/java/com/rogergcc/qrunlockclientapp/domain/model/AttendanceEntity.kt
@@ -1,0 +1,24 @@
+package com.rogergcc.qrunlockclientapp.domain.model
+
+import android.os.Parcelable
+import com.rogergcc.qrunlockclientapp.data.network.response.AttedancesRows
+
+
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class AttendanceEntity (
+        val id: Int?,
+        val chatId: String?,
+        val eventId: Int?,
+        val createdAt: String?,
+        val updatedAt: String?
+        ):Parcelable
+
+fun AttedancesRows.toDomain() = AttendanceEntity(
+        id,
+        chatId,
+        eventId,
+        createdAt,
+        updatedAt
+        )
