+++ app/src/main/java/com/rogergcc/qrunlockclientapp/presentation/helper/HoleView.kt
@@ -1,0 +1,60 @@
+package com.rogergcc.qrunlockclientapp.presentation.helper
+
+
+/**
+ * Created on junio.
+ * year 2023 .
+ */
+import android.content.Context
+import android.graphics.Canvas
+import android.util.AttributeSet
+import android.view.View
+
+class HoleView : View {
+    private var mDrawer: HoleDrawer? = null
+
+    constructor(context: Context) : super(context) {
+        init(context, null, 0, 0)
+    }
+
+    constructor(context: Context, attrs: AttributeSet?) : super(context, attrs) {
+        init(context, attrs, 0, 0)
+    }
+
+    constructor(context: Context, attrs: AttributeSet?, defStyleAttr: Int) : super(
+        context,
+        attrs,
+        defStyleAttr
+    ) {
+        init(context, attrs, defStyleAttr, 0)
+    }
+
+    constructor(
+        context: Context,
+        attrs: AttributeSet?,
+        defStyleAttr: Int,
+        defStyleRes: Int,
+    ) : super(context, attrs, defStyleAttr, defStyleRes) {
+        init(context, attrs, defStyleAttr, defStyleRes)
+    }
+
+    private fun init(context: Context, attrs: AttributeSet?, defStyleAttr: Int, defStyleRes: Int) {
+        setWillNotDraw(false)
+    }
+
+    fun setDrawer(drawer: HoleDrawer?) {
+        mDrawer = drawer
+        invalidate()
+    }
+
+    override fun onDraw(canvas: Canvas?) {
+        super.onDraw(canvas)
+        if (mDrawer != null) mDrawer!!.draw(
+            canvas,
+            width,
+            height,
+            width / 2,
+            height / 2
+        )
+    }
+}
