--- app/src/main/java/com/rogergcc/qrunlockclientapp/ui/helper/SoundPoolPlayer.kt
@@ -1,34 +1,0 @@
-package com.rogergcc.qrunlockclientapp.ui.helper
-
-import android.content.Context
-import android.media.AudioManager
-import android.media.SoundPool
-import android.os.Build
-import com.rogergcc.qrunlockclientapp.R
-
-
-/**
- * Created on agosto.
- * year 2022 .
- */
-open class SoundPoolPlayer(pContext: Context?) {
- private var mShortPlayer: SoundPool? = null
- private val mSounds: HashMap<Int, Int> = HashMap()
- fun playShortResource(piResource: Int) {
-  val iSoundId = mSounds[piResource] as Int
-  mShortPlayer!!.play(iSoundId, 0.99f, 0.99f, 0, 0, 1f)
- }
-
- fun release() {
-  mShortPlayer!!.release()
- }
-
- init {
-  mShortPlayer = SoundPool.Builder()
-   .setMaxStreams(4)
-   .build()
-
-//        this.mShortPlayer = new SoundPool(4, AudioManager.STREAM_MUSIC, 0);
-  mSounds[R.raw.bleep] = mShortPlayer!!.load(pContext, R.raw.bleep, 1)
- }
-}
